fix(core): keep heavy parsers and node-only module scope out of consumer bundles
The slimmed complement to the merged #19794 workerd slice (the edge
build itself now lives there, including this branch's grafted
getBuiltinModule post-process and workerd/edge-light conditions). What
remains is node/browser bundle hygiene at each defect's structural home:

- documents: pure naming helpers split into naming.ts (utils.ts
  re-exports them; all import sites keep working); mammoth and unpdf
  load dynamically when a docx/pdf is actually parsed, so no consumer
  bundle carries the parser toolchain.
- working-memory: readAttachmentAction imports the naming helpers from
  naming.ts and holds DocumentService as a type plus the literal service
  id, keeping the fs-heavy service module out of its graph.
- plugin-manager: repoRoot resolves on first handler run instead of
  process.cwd() at registry assembly; the exec/execFile promisifications
  guard on the builtin's presence and fail at use.

documents + working-memory + plugin-manager suites: 297 passed (297)
against post-#19794 develop.
